requirements.txt
imageio==2.37.0
lazy_loader==0.4
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyserial==3.5
pyFirmata==1.1.0
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
scikit-image==0.25.2
scipy==1.16.2
shiboken6==6.10.0
tifffile==2025.10.4
